@@ -14,7 +14,6 @@ dependencies:
   webview_flutter: ^4.7.0
   webview_flutter_web: ^0.2.2+2
   chopper: ^8.0.1+1
-  connectivity: ^3.0.6
   http: ^1.2.1
   equatable: ^2.0.3
   json_annotation: any
@@ -29,12 +28,12 @@ dependencies:
   firebase_messaging: ^15.0.2
   url_launcher: ^6.3.1
   url_launcher_ios: ^6.3.2
-  vibration: ^1.7.6
+  vibration: ^3.1.3
   firebase_analytics: ^11.1.0
   freezed: ^2.5.2
   freezed_annotation: ^2.4.1
-  slang: ^3.31.1
-  slang_flutter: ^3.31.0
+  slang: ^4.6.0
+  slang_flutter: ^4.6.0
   syncfusion_flutter_gauges: ^26.1.42
   flutter_svg: ^2.0.9
   package_info_plus: ^8.0.2
